requirements.txt
# requirements.txt
flask
python-dotenv
gunicorn
openai
PyGithub
requests